+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.0.2
+
+* Fixed issue when `wrap_content` width was breaking morphing animation
+
+## 1.0.1
+
+* Fixed issue when `app:colorIdle` attribute was ignored
+
 ## 1.0.0
 
 * Initial release